requirements.txt
flask
flask-socketio
scapy
eventlet
requests
python-socketio[client]
